refactor: expose useAuth hook via dedicated helper
