+++ src/pages/ProfileSettings.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { useData } from '../contexts/DataContext';
 import { motion } from 'framer-motion';
 import Navbar from '../components/layout/Navbar';
